Relocated some files
