+++ web-app/app/generator/AssignEmployeesStep.tsx
@@ -1,4 +1,3 @@
-// AssignEmployeesStep.tsx
 "use client"
 
 import { useMemo, useState } from "react"
