+++ web-app/app/generator/RuleDetailsStep.tsx
@@ -1,9 +1,9 @@
 "use client"
 
 import { useEffect } from "react"
-import { Button } from "./ui/button";
-import { Card } from "./ui/card";
-import { Input } from "./ui/input";
+import { Button } from "@/components/ui/button";
+import { Card } from "@/components/ui/card";
+import { Input } from "@/components/ui/input";
 import type { Rule, PriorityType } from "@/types"
 
 type Props = {
